// Labs/Assignment2/src/app/models/episode.ts
// this class is defined based on the structure of the JSON data returned from the API
export class Episode {
  id: number;
  url: string;
  name: string;
  season: number;
  number: number;
  airdate: string;
  airtime: string;
  airstamp: string;
  runtime: number;
  image: {
    medium: string;
    original: string;
  };
  summary: string;
  _links: {
    self: string;
  };
}