// Labs/Assignment2/src/app/models/show.ts
import { CastMember} from './castmember';
import { CrewMember } from './crewmember';
import { Episode } from './episode';

// this class is defined based on the structure of the JSON data returned from the API

export class Show {
  id: number;
  url: string;
  name: string;
  type: string;
  language: string;
  genres: string[];
  status: string;
  runtime: number;
  premiered: string;
  officialSite: string;
  schedule: {
    time: string;
    days: string[];
  };
  rating: {
    average: number;
  };
  weight: number;
  network: {
    id: number;
    name: string;
    country: {
      name: string;
      code: string;
      timezone: string;
    };
  };
  webChannel: {
    id: number;
    name: string;
    country: {
      name: string;
      code: string;
      timezone: string;
    };
  };
  image: {
    medium: string;
    original: string;
  };
  summary: string;
  updated: number;
  _links: {
    self: {
      href: string;
    };
    previousepisode: {
      href: string;
    };
  };
  _embedded: {
    episodes: Episode[];
    cast: CastMember[];
    crew: CrewMember[];
  };
}
